# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.13.0 🍺
 * Do not edit manually.
 * Api
 * AstroSentinel API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export type AstroEventEventType = typeof AstroEventEventType[keyof typeof AstroEventEventType];


export const AstroEventEventType = {
  GRB: 'GRB',
  GW: 'GW',
  FRB: 'FRB',
} as const;

/**
 * Normalised alert lifecycle state
 */
export type AstroEventLifecycle = typeof AstroEventLifecycle[keyof typeof AstroEventLifecycle];


export const AstroEventLifecycle = {
  preliminary: 'preliminary',
  initial: 'initial',
  update: 'update',
  confirmed: 'confirmed',
} as const;

/**
 * IceCube classification tier
 */
export type AstroEventClassificationTier = typeof AstroEventClassificationTier[keyof typeof AstroEventClassificationTier] | null;


export const AstroEventClassificationTier = {
  GOLD: 'GOLD',
  BRONZE: 'BRONZE',
} as const;

export interface AstroEvent {
  id: string;
  /** Human-readable event ID (e.g. GRB260503A) */
  eventId: string;
  eventType: AstroEventEventType;
  observatory: string;
  /** ISO timestamp with microsecond precision */
  detectionTime: string;
  /** Right Ascension in degrees */
  ra: number;
  /** Declination in degrees */
  dec: number;
  /** Error radius in arcminutes */
  errorRadius: number;
  /** Signal-to-noise ratio */
  snr: number;
  /** False alarm rate (Hz) */
  far: number;
  /** Fluence (erg/cm²), GRB only */
  fluence?: number | null;
  /** Dispersion measure (pc/cm³), FRB only */
  dm?: number | null;
  /** T90 duration (s), GRB only */
  t90?: number | null;
  /** Peak Flux, FRB and GRB */
  peakFlux?: number | null;
  /** Chirp mass (solar masses), GW only */
  chirpMass?: number | null;
  /** Luminosity distance (Mpc), GW only */
  luminosityDistance?: number | null;
  /** Galactic latitude */
  galLat: number;
  /** Galactic longitude */
  galLon: number;
  /** Angular distance from Sun (degrees) */
  sunDistance: number;
  /** Angular distance from Moon (degrees) */
  moonDistance: number;
  /** System latency in microseconds */
  latencyUs: number;
  createdAt: string;
  /** Normalised alert lifecycle state */
  lifecycle: AstroEventLifecycle;
  /** Raw alert_type string from originating source (e.g. PRELIMINARY, RETRACTION) */
  alertType?: string | null;
  /** IceCube classification tier */
  classificationTier?: AstroEventClassificationTier;
}

export interface EventListResponse {
  events: AstroEvent[];
  total: number;
}

export type EventStatsByType = {
  GRB: number;
  GW: number;
  FRB: number;
};

export type EventStatsByObservatoryItem = {
  observatory: string;
  count: number;
};

export interface EventStats {
  totalEvents: number;
  byType: EventStatsByType;
  byObservatory: EventStatsByObservatoryItem[];
  /** Events per hour in the last hour */
  recentRate: number;
  latestEvent?: AstroEvent | null;
}

export type ListEventsParams = {
limit?: number;
offset?: number;
eventType?: ListEventsEventType;
observatory?: string;
};

export type ListEventsEventType = typeof ListEventsEventType[keyof typeof ListEventsEventType];


export const ListEventsEventType = {
  GRB: 'GRB',
  GW: 'GW',
  FRB: 'FRB',
} as const;